Condensed tests and changed variable name.
